requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0

# Microsoft Agent Framework
agent-framework --pre
azure-identity>=1.15.0

# Stock market data
yfinance>=0.2.18

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Development
ipython>=8.12.0